CR#21136 Moved some packages
