+++ agent/src/main/java/com/teamscale/jacoco/agent/commit_resolution/sapnwdi/NwdiManifestException.java
@@ -1,4 +1,4 @@
-package com.teamscale.jacoco.agent.sapnwdi;
+package com.teamscale.jacoco.agent.commit_resolution.sapnwdi;
 
 /** Thrown in case an NWDI Manifest file cannot be read. */
 public class NwdiManifestException extends Exception {
