+++ agent/src/test/resources/com/teamscale/jacoco/agent/commit_resolution/git_properties/plain-git-properties.jar
