+++ agent/src/test/resources/com/teamscale/jacoco/agent/commit_resolution/git_properties/spring-boot-git-properties.war
